Make share menu smaller
